requirements.txt
rich~=12.3.0
python-dotenv~=0.20.0